+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #### Next
 
+* [#4](https://github.com/dblock/gem-licenses/pull/4): Try to match file contents to known, common licenses - [@joeyAghion](https://github.com/joeyAghion).
 * Added `gem:licenses` Rake task - [@dblock](https://github.com/dblock).
 * Attempt to normalize license name - [@dblock](https://github.com/dblock).
 * Install Rake tasks via `Gem::Licenses.install_tasks` - [@dblock](https://github.com/dblock).